35 clean up subpanel logic, fix bug where subpanel wouldnt position properly once left panel was closed
