+++ src/couldBeSharedComponents/LayoutPanels/PanelContentsWrapperWithOptionalSubpanel/PanelContentsWrapperWithOptionalSubpanel.module.scss
@@ -1,17 +1,15 @@
-.panelContentsWrapper {
-  border: dashed thin green;
+.subpanelPositionedParent {
   height: 100%;
-  position: relative;
+  position: relative; // to support absolute positioning of subpanel
 }
 .subpanelWrapper {
   border: solid thin cyan;
   height: 100%;
   width: 200px;
   position: absolute;
-  right: calc(
-    -100% - var(--spk-spacing-5, 24px)
-  ); // make sure changes here get reflected in component code for widthStylesOverride
+  right: 0;
   background: white;
   box-sizing: border-box;
   z-index: 1; // ensures stacked on top of map (tested with maplibre)
+  transform: translateX(calc(100% + var(--spk-spacing-5, 24px)));
 }
